move steps under ai folder
